moved logo to core
