+++ src/app/core/layout/logo/Logo.js
