+++ src/app/core/layout/logo/index.js
